pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-forge"
version = "0.1.0"
description = "Production-grade RAG system for business/legal documents"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "litellm>=1.0",
    "pymupdf4llm>=0.0.17",
    "python-docx>=1.1",
    "langchain-text-splitters>=0.3,<0.4",
    "numpy<2",
    "openai>=1.0",
    "chromadb>=0.5",
    "structlog>=24.0",
    "jinja2>=3.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
advanced = [
    "sentence-transformers>=2.2,<3.0",
    "onnxruntime<1.20.0; sys_platform == 'darwin' and platform_machine == 'x86_64'",
]
retrieval = [
    "cohere>=5.0",
    "flashrank>=0.2",
    "rank-bm25>=0.2",
]
eval = [
    "ragas>=0.2",
    "deepeval>=1.0",
]
observability = [
    "arize-phoenix>=4.0",
    "opentelemetry-sdk>=1.0",
]
api = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
]
cloud = [
    "qdrant-client>=1.9",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ipykernel>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rag_forge"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]